Refactor/rename '/commons/Constants.jsx' -> '/eMIB/constants.js'; correctign all relevant imports
