+++ frontend/src/components/eMIB/Inbox.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import EmailPreview from "./EmailPreview";
 import Email from "./Email";
 import "../../css/inbox.css";
-import { HEADER_HEIGHT, FOOTER_HEIGHT } from "../commons/Constants";
+import { HEADER_HEIGHT, FOOTER_HEIGHT } from "./constants";
 
 const INBOX_HEIGHT = `calc(100vh - ${HEADER_HEIGHT + FOOTER_HEIGHT}px)`;
